refactor(common): move interpolate-vars to bruno-common
