+++ packages/bruno-cli/src/runner/run-single-request.js
@@ -4,7 +4,7 @@ const decomment = require('decomment');
 const fs = require('fs');
 const { forOwn, each, extend, get, compact } = require('lodash');
 const prepareRequest = require('./prepare-request');
-const interpolateVars = require('./interpolate-vars');
+const { interpolateVars } = require('@usebruno/common');
 const { interpolateString, interpolateObject } = require('./interpolate-string');
 const { ScriptRuntime, TestRuntime, VarsRuntime, AssertRuntime, formatErrorWithContext, SCRIPT_TYPES } = require('@usebruno/js');
 const { stripExtension } = require('../utils/filesystem');
