+++ packages/bruno-common/src/index.ts
@@ -3,5 +3,6 @@ export { default as interpolate, interpolateObject } from './interpolate';
 export { percentageToZoomLevel } from './zoom';
 export { default as isRequestTagsIncluded } from './tags';
 export { transformExampleStatusInCollection } from './example-status';
+export { default as interpolateVars } from './interpolate-vars';
 
 export * as utils from './utils';
